pyproject.toml
[project]
name = "flooding-ai-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "langchain>=1.2.12",
    "langchain-openai>=1.1.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "spotipy>=2.25.1",
    "uvicorn>=0.42.0",
    "google-api-python-client>=2.70.0",
    "langchain-community>=0.4.1",
    "pymupdf>=1.27.2.2",
    "langchain-chroma>=1.1.0",
    "langchain-core>=1.2.23",
    "langgraph>=1.1.3",
]
